Update the number of remaining unprocessed bytes correctly
Update uio structure in sock_net_write/read. VFSCORE expects the
update of the resid member of uio such that it reflects the
number of bytes that remain unprocessed.

Set the UK_VFSCORE_NOPOS flag on the socket vfscore_file. This
flag informs VFSCORE to do not update the offset of our socket
file since it doesn't make sense for sockets.

Signed-off-by: Mihai Pogonaru <pogonarumihai@gmail.com>
Reviewed-by: Costin Lupu <costin.lupu@cs.pub.ro>