Correct license headers introduced by NEC Laboratories Europe GmbH
This commit removes the additional clause
"THIS HEADER MAY NOT BE EXTRACTED OR MODIFIED IN ANY WAY"
from BSD license headers that got introduced by mistake
with the following commits:

 7d8209c - Add prototype for libvfscore for Sockets
 cf5ec19 - Update to lwIP 2.1.2
 96d4f49 - Initial port of lwIP to Unikraft
 d01ed1c - Netif driver for libuknetdev devices

On the one hand, this additional clause is redundant because
the BSD license already states that it must remain and the
copyright notice must be kept.

On the other hand, the clause freezes the file header
and prohibits future contributors from extending the
copyright notice for their contributions. This additional
clause is not part of the official BSD 3-clause.

The original author or affiliation consents to the change
with a `Signed-off-by` or `Reviewed-by` tag to this
commit.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Signed-off-by: Sharan Santhanam <sharan.santhanam@neclab.eu>
Signed-off-by: Felipe Huici <felipe.huici@neclab.eu>