uknetdev: Enable support for hardware checksum offloading
This commit implements and enables checksum offloading with
the Unikraft fork of lwIP and if the underlying uknetdev
device announced support for it. For non-supported devices,
all checksums are checked and computed. When an lwIP stack
version is used that currently does not support checksum
offloading, the original behavior is kept.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>