Option to choose latest lwIP version from 2.1.x series
Introduces an option to choose latest lwIP version from
the 2.1.x series. This version is based on the latest
official lwIP release and includes further bugfixes and
security fixes. Because we point to a branch head,
recent commits could break compiling. Because of this,
we mark this option as `unofficial`.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>