Option for latest lwIP 2.1.x version including Unikraft patches
Introduces an option to choose the latest lwIP version from
the 2.1.x series that is tested by the Unikraft team. This
version is based on the latest official lwIP release,
includes some bugfixes and security fixes, and includes
feature enhancements done by the Unikraft community. These
enhancements are not yet upstream available as part of the
official lwIP project (upstreaming ongoing).

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>