Add mailbox configuration options
This fixes a bug where some packets are dropped because
LWIP mailbox is bigger than the UDP and TCP mailboxes.

This introduces the "LWIP stackthread mbox size" where you
can now select the size for the input mailbox for the lwip thread.

The sizes for the UDP and TCP mailboxes are now always bigger or
equal to the LWIP input mailbox. You can select the factor by which
they are greater in the "Sockets API" submenu. (i.e if you choose
a factor of 2 for the UDP mailbox it will have the size 2*LWIP mailbox
size).

Signed-off-by: Daniel Dinca <danieldinca97@yahoo.com>