Remove redundant argument from POSIX_SOCKET_FAMILY_REGISTER macro
In the `POSIX_SOCKET_FAMILY_REGISTER` macro, when the IPv6 option is
selected, the third argument produces a compilation error.

Due to the macro being defined in
`unikraft/lib/posix-socket/include/uk/socket_driver.h` with only 2
arguments, the third argument becomes redundant.

This commit removes it and ensures that the `AF_INET6` is successfully
registered.

Signed-off-by: Maria Sfiraiala <maria.sfiraiala@gmail.com>