Fix packet leak on uknetdev_output()
uk_netdev_tx_one() returns at least UK_NETDEV_STATUS_SUCESS if the packet was
consumed. It might still return 0 if there was no space in the transmit ring, in
that case the packet is not consumed and it's the caller's responsability to
deal with it.

Observe virtio_net.c:445, virtio_netdev_xmit():

	} else if (rc == -ENOSPC) {
		uk_pr_debug("No more descriptor available\n");
		/**
		 * Remove header before exiting because we could not send
		 */
		uk_netbuf_header(pkt, -header_sz);
...
	return (status) /* which is zero here */

I find the API a bit confusing, it would make more sense to not return zero from
netdev, but it's properly documented:

 * @return
 *   - (>=0): Positive value with status flags
 *     - UK_NETDEV_STATUS_SUCCESS: `pkt` was successfully put to the transmit
 *        queue. Whenever this flag is not set, there was no space left on the
 *        transmit queue to send `pkt`.

Untested, caught just by reading the virtio driver, hope I'm not missing anything.